requirements.txt
requests
pyyaml
PyQt5
python-dotenv
pytest
allure-pytest